Update tests to remove calls to resultFunc
Using resultFunc means we don't test whether the correct parent selectors are called. These could be changed, breaking functionality, but without breaking the tests. As such it's better to test the full selector. This also behaves like a proper black box test where we pass the true expected input and test the true returned output. The alternative would be to assert against "selector.dependencies", but this is currently not in the type definitions and so throws an error when used.